Setup Qt 5.14 github workflow

+++ src/core/expression/qgsexpressionfunction.cpp
@@ -404,7 +404,7 @@ static QVariant fcnRndF( const QVariantList &values, const QgsExpressionContext 
   }
 
   // Return a random double in the range [min, max] (inclusive)
-  double f = static_cast< double >( generator() ) / generator.max();
+  double f = static_cast< double >( generator() ) / static_cast< double >( generator.max() );
   return QVariant( min + f * ( max - min ) );
 }
 static QVariant fcnRnd( const QVariantList &values, const QgsExpressionContext *, QgsExpression *parent, const QgsExpressionNodeFunction * )
